+++ api/src/main/java/com/polikt/api/agency/AgencyRepository.java
@@ -2,6 +2,6 @@ package com.polikt.api.agency;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgencyRepository extends JpaRepository<Agency, Integer> {
+public interface AgencyRepository extends JpaRepository<Agency, Long> {
 
 }
