+++ api/src/main/java/com/polikt/api/guide/GuideController.java
@@ -1,15 +1,15 @@
 package com.polikt.api.guide;
 
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import java.util.List;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-
-import java.util.List;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/guides")
@@ -27,7 +27,7 @@ public class GuideController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Guide> getGuideById(@PathVariable int id) {
+    public ResponseEntity<Guide> getGuideById(@PathVariable Long id) {
         Guide guide = repository.findById(id).orElse(null);
 
         if (guide != null) {
@@ -43,7 +43,7 @@ public class GuideController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteGuideById(@PathVariable int id) {
+    public ResponseEntity<Void> deleteGuideById(@PathVariable Long id) {
         if (!repository.existsById(id)) {
             return ResponseEntity.notFound().build();
         }
